refactor(#737): eliminate Client from E2EE widgets and relocate controllers
- Relocate 4 controller/handler files from widgets/ to services/:
  bootstrap_controller, bootstrap_driver, key_backup_signer,
  recovery_key_handler (all are non-widget logic, belong in services/)
- Update all import paths in lib/ and test/
- Remove matrix import from e2ee_setup_screen.dart:
  replace StreamSubscription<UiaRequest<dynamic>> with
  passwordPromptBuilder callback (same pattern as devices_screen)
- flutter analyze: 0 issues, flutter test: 2318 pass
