+++ lib/core/services/sub_services/chat_backup_service.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:kohera/core/services/sub_services/backup_version_manager.dart';
-import 'package:kohera/features/e2ee/widgets/key_backup_signer.dart';
+import 'package:kohera/features/e2ee/services/key_backup_signer.dart';
 import 'package:matrix/encryption.dart';
 import 'package:matrix/matrix.dart';
