+++ lib/features/e2ee/services/key_backup_signer.dart
