+++ lib/features/e2ee/services/recovery_key_handler.dart
