+++ test/widgets/key_backup_signer_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:kohera/features/e2ee/widgets/key_backup_signer.dart';
+import 'package:kohera/features/e2ee/services/key_backup_signer.dart';
 import 'package:matrix/encryption.dart';
 import 'package:matrix/encryption/olm_manager.dart';
 import 'package:matrix/matrix.dart';
